requirements.txt
piexif
Pillow
argparse
gpxpy
pytz
timezonefinder
